+++ pkg/alertmanager/alertmanager.go
@@ -9,11 +9,9 @@ import (
 	"net/url"
 	"time"
 
-	"github.com/pkg/errors"
-
-	"github.com/giantswarm/silence-operator/api/v1alpha1"
-	"github.com/giantswarm/silence-operator/api/v1alpha2"
 	"github.com/giantswarm/silence-operator/pkg/config"
+	"github.com/pkg/errors"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 const (
@@ -31,7 +29,20 @@ var (
 	ErrSilenceNotFound = errors.New("silence not found")
 )
 
-type AlertManager struct {
+// Client defines the contract for alertmanager operations
+type Client interface {
+	GetSilenceByComment(comment string) (*Silence, error)
+	CreateSilence(s *Silence) error
+	UpdateSilence(s *Silence) error
+	DeleteSilenceByComment(comment string) error
+	DeleteSilenceByID(id string) error
+	ListSilences() ([]Silence, error)
+}
+
+// Ensure Alertmanager implements Client
+var _ Client = (*Alertmanager)(nil)
+
+type Alertmanager struct {
 	address        string
 	authentication bool
 	token          string
@@ -39,12 +50,12 @@ type AlertManager struct {
 	client         *http.Client
 }
 
-func New(config config.Config) (*AlertManager, error) {
+func New(config config.Config) (*Alertmanager, error) {
 	if config.Address == "" {
 		return nil, errors.Errorf("%T.Address must not be empty", config)
 	}
 
-	return &AlertManager{
+	return &Alertmanager{
 		address:        config.Address,
 		authentication: config.Authentication,
 		token:          config.BearerToken,
@@ -53,7 +64,7 @@ func New(config config.Config) (*AlertManager, error) {
 	}, nil
 }
 
-func (am *AlertManager) GetSilenceByComment(comment string) (*Silence, error) {
+func (am *Alertmanager) GetSilenceByComment(comment string) (*Silence, error) {
 	silences, err := am.ListSilences()
 	if err != nil {
 		return nil, errors.WithStack(err)
@@ -68,7 +79,7 @@ func (am *AlertManager) GetSilenceByComment(comment string) (*Silence, error) {
 	return nil, errors.WithMessagef(ErrSilenceNotFound, "failed to get silence with comment %#q", comment)
 }
 
-func (am *AlertManager) CreateSilence(s *Silence) error {
+func (am *Alertmanager) CreateSilence(s *Silence) error {
 	endpoint := fmt.Sprintf("%s%s", am.address, apiV2SilencesPath) // Use constant
 
 	jsonValues, err := json.Marshal(s)
@@ -99,14 +110,14 @@ func (am *AlertManager) CreateSilence(s *Silence) error {
 	return nil
 }
 
-func (am *AlertManager) UpdateSilence(s *Silence) error {
+func (am *Alertmanager) UpdateSilence(s *Silence) error {
 	if s.ID == "" {
 		return errors.Errorf("failed to update silence %#q, missing ID", s.Comment)
 	}
 	return am.CreateSilence(s)
 }
 
-func (am *AlertManager) DeleteSilenceByComment(comment string) error {
+func (am *Alertmanager) DeleteSilenceByComment(comment string) error {
 	silences, err := am.ListSilences()
 	if err != nil {
 		return errors.WithStack(err)
@@ -121,7 +132,7 @@ func (am *AlertManager) DeleteSilenceByComment(comment string) error {
 	return errors.WithMessagef(ErrSilenceNotFound, "failed to delete silence by comment %#q", comment)
 }
 
-func (am *AlertManager) ListSilences() ([]Silence, error) {
+func (am *Alertmanager) ListSilences() ([]Silence, error) {
 	endpoint := fmt.Sprintf("%s%s", am.address, apiV2SilencesPath)
 
 	var silences []Silence
@@ -161,7 +172,7 @@ func (am *AlertManager) ListSilences() ([]Silence, error) {
 	return filteredSilences, nil
 }
 
-func (am *AlertManager) DeleteSilenceByID(id string) error {
+func (am *Alertmanager) DeleteSilenceByID(id string) error {
 	// Use constant and url.PathEscape for safety if ID can contain special chars
 	endpoint := fmt.Sprintf("%s%s/%s", am.address, apiV2SilencePath, url.PathEscape(id))
 
@@ -189,19 +200,18 @@ func (am *AlertManager) DeleteSilenceByID(id string) error {
 	return nil
 }
 
-func SilenceComment(silence *v1alpha1.Silence) string {
-	return fmt.Sprintf("%s-%s", CreatedBy, silence.Name)
-}
-
-func SilenceCommentV1Alpha2(silence *v1alpha2.Silence) string {
-	return fmt.Sprintf("%s-%s-%s", CreatedBy, silence.Namespace, silence.Name)
+func SilenceComment(silence client.Object) string {
+	if silence.GetNamespace() != "" {
+		return fmt.Sprintf("%s-%s-%s", CreatedBy, silence.GetNamespace(), silence.GetName())
+	}
+	return fmt.Sprintf("%s-%s", CreatedBy, silence.GetName())
 }
 
 // SilenceEndsAt gets the expiry date for a given silence.
 // The expiry date is retrieved from the annotation name configured by ValidUntilAnnotationName.
 // The expected format is defined by DateOnlyLayout.
 // It returns an invalidExpirationDateError in case the date format is invalid.
-func SilenceEndsAt(silence *v1alpha1.Silence) (time.Time, error) {
+func SilenceEndsAt(silence client.Object) (time.Time, error) {
 	annotations := silence.GetAnnotations()
 
 	// Check if the annotation exist otherwise return a date 100 years in the future.