+++ pkg/alertmanager/http.go
@@ -7,7 +7,7 @@ import (
 
 // NewRequest creates a new http.Request with the given method, url and body.
 // It adds the tenantId as X-Scope-OrgID header to the request if it is set.
-func (am *AlertManager) NewRequest(method, url string, body io.Reader) (*http.Request, error) {
+func (am *Alertmanager) NewRequest(method, url string, body io.Reader) (*http.Request, error) {
 	req, err := http.NewRequest(method, url, body)
 	if err != nil {
 		return nil, err